+++ src/Services/Sale/SaleServiceBuilder.php
@@ -22,11 +22,21 @@ use Bitrix24\SDK\Services\Sale\PropertyVariant\Service\PropertyVariant;
 use Bitrix24\SDK\Services\Sale\Status\Service\Status;
 use Bitrix24\SDK\Services\Sale\StatusLang\Service\StatusLang;
 use Bitrix24\SDK\Services\Sale\PersonTypeStatus\Service\PersonTypeStatus;
+use Bitrix24\SDK\Services\Sale\ShipmentPropertyValue\Service\ShipmentPropertyValue;
+use Bitrix24\SDK\Services\Sale\ShipmentItem\Service\ShipmentItem;
 use Bitrix24\SDK\Services\Sale\BasketProperty\Service\BasketProperty;
 use Bitrix24\SDK\Services\Sale\DeliveryHandler\Service\DeliveryHandler;
 use Bitrix24\SDK\Services\Sale\Delivery\Service\Delivery;
 use Bitrix24\SDK\Services\Sale\DeliveryRequest\Service\DeliveryRequest;
 use Bitrix24\SDK\Services\Sale\DeliveryExtraService\Service\DeliveryExtraService;
+use Bitrix24\SDK\Services\Sale\Payment;
+use Bitrix24\SDK\Services\Sale\PaymentItemBasket;
+use Bitrix24\SDK\Services\Sale\PaymentItemShipment;
+use Bitrix24\SDK\Services\Sale\PersonType;
+use Bitrix24\SDK\Services\Sale\PropertyGroup;
+use Bitrix24\SDK\Services\Sale\Order;
+use Bitrix24\SDK\Services\Sale\BasketItem;
+use Bitrix24\SDK\Services\Sale\PropertyRelation;
 
 /**
  * Class SaleServiceBuilder
@@ -149,6 +159,36 @@ class SaleServiceBuilder extends AbstractServiceBuilder
         return $this->serviceCache[__METHOD__];
     }
 
+    /**
+     * Payment Item Basket service (sale.paymentitembasket.*)
+     */
+    public function paymentItemBasket(): PaymentItemBasket\Service\PaymentItemBasket
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new PaymentItemBasket\Service\PaymentItemBasket(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    /**
+     * Payment Item Shipment service (sale.paymentitemshipment.*)
+     */
+    public function paymentItemShipment(): PaymentItemShipment\Service\PaymentItemShipment
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new PaymentItemShipment\Service\PaymentItemShipment(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
     public function propertyVariant(): PropertyVariant
     {
         if (!isset($this->serviceCache[__METHOD__])) {
@@ -174,6 +214,36 @@ class SaleServiceBuilder extends AbstractServiceBuilder
     }
 
     /**
+     * Shipment service (sale.shipment.*)
+     */
+    public function shipment(): Shipment\Service\Shipment
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Shipment\Service\Shipment(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    /**
+     * ShipmentProperty service (sale.shipmentproperty.*)
+     */
+    public function shipmentProperty(): ShipmentProperty\Service\ShipmentProperty
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new ShipmentProperty\Service\ShipmentProperty(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    /**
      * Basket Item service (sale.basketitem.*)
      */
     public function basketItem(): BasketItem\Service\BasketItem
@@ -185,6 +255,36 @@ class SaleServiceBuilder extends AbstractServiceBuilder
             );
             $this->serviceCache[__METHOD__] = new BasketItem\Service\BasketItem(
                 new BasketItem\Service\Batch($batch, $this->log),
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    /**
+     * ShipmentPropertyValue service (sale.shipmentpropertyvalue.*)
+     */
+    public function shipmentPropertyValue(): ShipmentPropertyValue
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new ShipmentPropertyValue(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    /**
+     * ShipmentItem service (sale.shipmentitem.*)
+     */
+    public function shipmentItem(): ShipmentItem
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new ShipmentItem(
                 $this->core,
                 $this->log
             );
@@ -260,6 +360,21 @@ class SaleServiceBuilder extends AbstractServiceBuilder
     {
         if (!isset($this->serviceCache[__METHOD__])) {
             $this->serviceCache[__METHOD__] = new DeliveryExtraService(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+  
+    /**
+     * Property Relation service (sale.propertyRelation.*)
+     */
+    public function propertyRelation(): PropertyRelation\Service\PropertyRelation
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new PropertyRelation\Service\PropertyRelation(
                 $this->core,
                 $this->log
             );